pfarrell/bemused#34: Extract service layers for auth, streams, albums, search routes
